Setup Foundry, Add Integration Test, Deploy Script (#7)
This initializes a new foundry project in this repository. In addition to adding the base foundry project files, an integration test is added for testing the deployment and interaction on a forked Optimism network.

+++ contracts/external/weth/IWETH.sol
@@ -2,11 +2,11 @@
 pragma solidity ^0.8.0;
 
 interface IWETH {
-    function deposit() external payable;
+  function deposit() external payable;
 
-    function transfer(address to, uint value) external returns (bool);
+  function transfer(address to, uint256 value) external returns (bool);
 
-    function withdraw(uint) external;
+  function withdraw(uint256) external;
 
-    function balanceOf(address account) external view returns (uint256);
+  function balanceOf(address account) external view returns (uint256);
 }
